ui cleanup

+++ quick-start/src/main/ui/app/login/login.component.ts
@@ -1,14 +1,13 @@
 import { Component, Input, Renderer, trigger,
    state, style, transition, animate } from '@angular/core';
 import { Router } from '@angular/router';
-import { STOMPService } from '../stomp/stomp.service';
 
 import * as _ from 'lodash';
 
 import { AuthService } from '../auth/auth.service';
 import { ProjectService } from '../projects/projects.service';
-import { InstallService } from '../installer/install.service';
-
+import { InstallService } from '../installer';
+import { ConfirmService } from '../confirm';
 import { FolderBrowser } from '../folder-browser/folder-browser.component';
 import { SelectList } from '../select-list/select-list.component';
 import { LoginInfo } from './login-info.model';
@@ -21,7 +20,7 @@ import { HubSettings } from './hub-settings.model';
     FolderBrowser,
     SelectList
   ],
-  providers: [AuthService, ProjectService, InstallService, STOMPService],
+  providers: [],
   styleUrls: ['./login.style.scss'],
   animations: [
     trigger('visibleState', [
@@ -59,7 +58,7 @@ export class Login {
   };
 
 
-  progressHidden = true;
+  installing: boolean = false;
 
   environments = ['local', 'dev', 'qa', 'prod'];
 
@@ -89,9 +88,11 @@ export class Login {
     private auth: AuthService,
     private projectService: ProjectService,
     private installService: InstallService,
+    private confirm: ConfirmService,
     private router: Router,
     private renderer: Renderer) {
 
+    auth.setAuthenticated(false);
     projectService.getProjects().subscribe(resp => {
       this.projects = resp.projects;
 
@@ -110,7 +111,7 @@ export class Login {
   }
 
   install() {
-    this.progressHidden = false;
+    this.installing = true;
 
     this.installationStatus = '';
     this.installService.messageEmitter.subscribe((payload) => {
@@ -119,11 +120,18 @@ export class Login {
 
       if (this.percentComplete === 100) {
         setTimeout(() => {
-          this.progressHidden = true;
+          this.installing = false;
         }, 1000);
+        this.currentEnvironment.installed = true;
       }
     });
-    this.installService.install();
+    this.installService.install(
+      this.currentProject.id,
+      this.currentEnvironmentString);
+  }
+
+  installNext() {
+    this.router.navigate(['']);
   }
 
   folderClicked(folder) {
@@ -165,12 +173,19 @@ export class Login {
   }
 
   removeProject(project) {
-    this.projectService.removeProject(project).subscribe(() => {
-      _.remove(this.projects, p => { return p.id === project.id; });
-      if (this.projects.length === 0) {
-        this.showFolderBrowser = true;
-      }
-    });
+    this.confirm.showConfirm({
+      title: 'Remove Project?',
+      message: 'Remove the project from the list of projects? Does not destroy anything on disk.',
+      okText: 'Remove',
+      cancelText: 'Cancel'
+    }).then(() => {
+      this.projectService.removeProject(project).subscribe(() => {
+        _.remove(this.projects, p => { return p.id === project.id; });
+        if (this.projects.length === 0) {
+          this.showFolderBrowser = true;
+        }
+      });
+    }).catch(() => {});
   }
 
   gotProject = (project) => {
@@ -189,14 +204,14 @@ export class Login {
   }
 
   restoreInitDefaults() {
-    // const confirm = this.$mdDialog.confirm()
-    //       .title('Are you sure?')
-    //       .textContent('Really restore the default settings?')
-    //       .ok('Restore')
-    //       .cancel('Cancel');
-    // this.$mdDialog.show(confirm).then(() => {
+    this.confirm.showConfirm({
+      title: 'Are you sure?',
+      message: 'Really restore the default settings?',
+      okText: 'Restore',
+      cancelText: 'Cancel'
+    }).then(() => {
       this.initSettings = _.clone(this.defaultSettings);
-    // });
+    }).catch(() => {});
   }
 
   gotEnvironment(environment) {
@@ -221,7 +236,7 @@ export class Login {
 
       if (this.currentEnvironment.installed) {
         // goto login tab
-        this.router.navigate(['home']);
+        this.router.navigate(['']);
       } else {
         // go to install hub tab
         this.gotoTab('Installer');
